feat: refrige, user 파일 폴더링
