--- backend/src/main/java/still88/backend/domain/user/UserController.java
@@ -1,23 +1,0 @@
-package still88.backend.domain.user;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.ResponseEntity;
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import still88.backend.dto.user.UserRequestDto;
-
-@Controller
-@RequiredArgsConstructor
-@RequestMapping("/user")
-public class UserController {
-
-    private final UserService userService;
-
-    // 유저 생성하기
-    @PostMapping("/create")
-    public ResponseEntity<?> insertUser(@RequestBody UserRequestDto userRequestDto) {
-        return ResponseEntity.ok(userService.createUser(userRequestDto));
-    }
-}
