+++ backend/src/main/java/still88/backend/domain/user/service/UserService.java
@@ -1,4 +1,8 @@
-package still88.backend.domain.user.service;
+package still88.backend.domain.user;
 
-public class UserService {
+import still88.backend.dto.user.UserRequestDto;
+import still88.backend.dto.user.UserResponseDto;
+
+public interface UserService {
+    public UserResponseDto createUser(UserRequestDto userRequestDto);
 }
